Discussed changes
I moved config files for UserDefaults into a subdir to make it more apparent that they are config. I added a default delay time as an extension to UserDefaults too and replaced hard coded times in the code base. DocumentGroup extensions onboardingModal and splashScreenModal were not functioning, so I made small changes to get them working. Swept docs for cleanup.